+++ NEWS.md
@@ -1,3 +1,15 @@
+# matrixset 0.4.1
+
+## Internal changes
+
+Patch to comply with
+
+* tidyverse/tidyr#1622
+* Matrix deprecating the use of as(., "dgeMatrix")
+
+The documentation no longer uses the pipe (|>) operator, to avoid dependency on
+R >= 4.1.0.
+
 # matrixset 0.4.0
 
 ## Changes to `apply_matrix()`, `apply_row()` and `apply_column()`